+++ app/src/components/Onboarding/index.ts
@@ -1,0 +1,2 @@
+export { default as OnboardingView } from "./OnboardingView";
+export { default as DependencyCheck } from "./DependencyCheck";